okd-4.1: avoid permission denied running as user (#153)
If we do `make cluster-up` as user, almost at the end of
the procedure we get "permission denied error", because the
code tries to copy root-owned `kubectl` and `kubeconfig` from
the guest cluster to the host.

The copy succeeds, but just after the code also tries to make
sure `kubectl` is executable, a step which seems redundant on its own.

The copies are done in the local tree, so to keep the files
owned by root is incorrect: the should be owned by the user(:group)
running the actual command.

To overcome both these issues we add a helper function to mimic
the `install` command, copying a file from the cluster into the
local tree and setting the permissions and the owner(:group) in
one go.

Internally, we implement the transfer with a simple `cat`, just
because it is the simplest approach - and the first tried which worked.

Signed-off-by: Francesco Romani <fromani@redhat.com>