cluster: add support for different runtimes (#168)
* cluster: add support for different runtimes

with this patch, we open the road for different helper tools in addition
to the standard `gocli`, which is still the default choice.
We want to make room for the future podman support, which is currently
experimented with in `github.com/fromanirh/pack8s`, and this seems
the safest and less invasive way.

Signed-off-by: Francesco Romani <fromani@redhat.com>

* cluster: move container var under conditional

We use the _cli_container var iff we want to use
the stable `gocli` tool, hence move the variable declaration
under the if branch.

Signed-off-by: Francesco Romani <fromani@redhat.com>