okd-provider: rotate CA cert on startup
Since we use static images for the providers, we get a cert during
the provisioning of the provider that expires after 30 days.

With this change, every time we run the cluster, we force openshift's
CA service to rotate the certs so that we get a valid CA to work with
during our tests.

Signed-off-by: Daniel Belenky <dbelenky@redhat.com>