Activate serial pty qemu option to attach serial to PTY, add to new images.
This commit applying the change from #152 PR to new images and providers.

Adding this change also to the vm.sh script which copied to the base image,
making this functionalty avilable on new providers and images.

Add extra arg to the qemu exec command (creating qemu vm) '-serial pty'.
In case we loose connectivity to the vm we could still connect via 'socat - /dev/pts/0'
from the host.

Signed-off-by: Or Mergi <omergi@redhat.com>