Configure calico additional settings to work with NetworkManager (#281)
* Add config-cni.sh script to common-scripts

This script configuring additional settings to cni plugin,
and should run on provision stage.

If in the future we will need to configure another setting to
cni plugins we can simply add it to the correspondig function.

Signed-off-by: Or Mergi  <omergi@redhat.com>

* Run configure-cni.sh script at provision.sh

Source config-cni.sh and run in order to configure cni plugin
additional settings

Signed-off-by: Or Mergi  <omergi@redhat.com>

* Copy new script, config-cni.sh, to the VM

Signed-off-by: Or Mergi <omergi@redhat.com>

* Use conig-cni.sh in k8s-fedora provider provision.

Signed-off-by: Or Mergi <omergi@redhat.com>

* Update images hash.

Signed-off-by: Or Mergi <omergi@redhat.com>