kind: Don't relabel nodes with with schedule=true if not needed
When we startup a kind provider with just one node, it fails since master node is schedulable, and we cannot taint it with kubectl taint nodes $node node-role.kubernetes.io/master:NoSchedule-.

This PR label master nodes with kubevirt.io/schedulable=true only if we can taint the master nodes.

Also extra mile we have create a pair of helper functions to retrieve nodes and pods.

Signed-off-by: Quique Llorente <ellorent@redhat.com>