kind: fix SR-IOV 1.17 provider (#319)
* explicitly set SHELL of sriov-operator makefile to bash

We can't pass the SHELL settings to a makefile as a variable. Instead, we need
to set it as an argument of make. With this change, the SR-IOV lane should work
again.

Signed-off-by: Petr Horacek <phoracek@redhat.com>

* Move policy creation after configuration of webhook

With the current code, we try to apply an SriovNodeNetworkPolicy before
configuring webhook's secrets. This patch fixes the order.

Signed-off-by: Petr Horacek <phoracek@redhat.com>

* kind-sriov: make sure it is used only with two nodes

The script is only able to move SR-IOV device to the first worker node. Make
sure it is clear to users by returning an error if started without 2 nodes.

Signed-off-by: Petr Horacek <phoracek@redhat.com>

* remove sriov-1.14 lane

It does not work and we don't intend to fix it. It uses old Kubernetes, old
sriov-operator.

Signed-off-by: Petr Horacek <phoracek@redhat.com>

* worker nodes are already able

It is already done in the base config.

Signed-off-by: Petr Horacek <phoracek@redhat.com>

* kind-sriov: wait 60 seconds for certificate changes

30 was not enough in some cases.

Signed-off-by: Petr Horacek <phoracek@redhat.com>

* print NICs info in logs

Signed-off-by: Petr Horacek <phoracek@redhat.com>