kind: deploy local volume provisioner (#323)
Kind default StorageClass is named 'standard', while kubevirt
func tests assume it to be called 'local' - as is for all the
other providers.

To enable the kind-k8s-1.17-ipv6 provider to work out of the box,
the `local-volume-provisioner` daemonset must be deployed. It will
take care of:
  - create the 'local' storage class
  - discover all mount points in its configuration path (via config
    map) and expose those mount points as PersistentVolumes.

For more information, check the local static provisioner repo,
located at [0].

This will enable the functional_tests labeled as 'PVC' to be
executed.

[0] - https://github.com/kubernetes-sigs/sig-storage-local-static-provisioner

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>