dnsmasq: Create secondary networks
Create a network for each secondary NIC.
On each node, the secondary NIC X will be connected to network X
where the network is represented by a bridge (brX).

With this change, the secondary nics are no longer connected to the pod
network and therefore also no longer have a dynamic IP.

One of the known issues this change solves is the console/vnc access to
the KubeVirt VM/s when secondary nics are defined.

Includes updated base and cluster-up images hash.

Signed-off-by: Edward Haas <edwardh@redhat.com>