Bump the version of the kubevirt-testing image to fc32
The current version of the kubevirt-testing image is fc28; other
containers - like cdi-importer - are based on this image.

The cdi-importer exposes a set of binaries to the KubeVirt
functest VMs (qemu-ga, stress, etc); the issue here is that the
exposed binaries are simply installed via `dnf`. This means that
the version of the qemu-guest-agent being used on the functests
is a fc28 version.

The current fedora version used in the tests is fc32; this PR tries
to align the images of both the cdi-importer and the test VMs.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>