Fix support of secondary interfaces, K8s-1.17
NetworkManager-wait-online.service is causing the boot to take
much longer when secondary interfaces exists,
as a result, check-up during provision and cluster-up
would fail connecting to the node as long as the boot isnt finished.
By installing NetworkManager-config-server,
existing network profiles wont be changed, and the boot wont
be stalled by NetworkManager-wait-online.service.

Signed-off-by: Or Shoval <oshoval@redhat.com>