Add provision script to update kernel args for Vhostuser
The default kernel args as part of the provision.sh script has
64 2M hugepages. DPDK requires 1GB of hugepage to enable the
support and additionally VM also needs huagepages. And iommu
should be abled in order to used the PMD driver.

Kernel should be updated to the latest version for enabling DPDK.

All the above operations involes reboot of the node, so it is
grouped together for the DPDK requirement. This script can be
run as part of the provision itself, and it can be common to all
other nodes, but it causes memory overhead. Insted it is better
to invoke this script runing run gocli command which can be
followed up by a reboot, managed in the gocli itself.

This patch adds the script only, the invocation of this script
will be done in the follow-up patches.

Signed-off-by: Saravanan KR <skramaja@redhat.com>