Create OvS-DPDK operator when enabled
In order to create a VM with vhostuser interface, OvS-DPDK should
be enabled and relevant parameter should be configured.
ovsdpdk-network-operator does the configuration based on the
provided interface configuration. In order to use vhostuser
interface on VMs as additional interface, a secondary
interface should be added to the deployed nodes. And
operator config will specify the additional interface eth1
to configure it for vhostuser socket. Relevant manifests
are added to the provider to perform this operation.

Signed-off-by: Saravanan KR <skramaja@redhat.com>