Wait for 'NoSchedule' taint absence
After creating new SriovNetworkNodePolicy 'sriov-cni'
and 'sriov-device-plugin' pods are scheduled on the nodes
that match the policy, which cause a 'NoSchedule' taint on
that node that will remain untill those pods are up and ready.
Thus it is mandatory to wait for the taint to be removed
(after the mentioned pods are ready).

At 'wait_for_taint_absence' notice the '-v' flag at
the end of $action, propagates the string 'echo'ed
by is_taint_absence.

Signed-off-by: Or Mergi <ormergi@redhat.com>