Make base image ipv6 ready and use it for centos8 (#345)
* Make base image IPv6 ready

Update dnsmasq for ipv6 support.

Let dnsmasq advertise ipv6 addresses which can reach each other
over the internal br0 bridge in addition to the ipv4 addresses.

The k8s 1.17/1.18 clusters which are based on centos8, are configured to
automatically ask for an ipv6 address.
The IPv6 addresses will be assigned as follows:
node01: fd00::101
node02: fd00::102
and so forth.

Signed-off-by: Roman Mohr <rmohr@redhat.com>
Signed-off-by: Or Shoval <oshoval@redhat.com>

* Images: Update k8s-1.17 and k8s-1.18 Centos8 base image

Signed-off-by: Or Shoval <oshoval@redhat.com>

* Images: Update k8s-1.17 and k8s-1.18 hashes

Signed-off-by: Or Shoval <oshoval@redhat.com>

Co-authored-by: Roman Mohr <rmohr@redhat.com>