kind, automation: detach all loopbacks on provider startup
When running docker in docker (which is how kind does it's thing)
, the `losetup` command ends up looking for the next available
device in the host's namespace, instead of using the k8s node
namespace.

Eventually, it will grow past the number of statically allocated
loopback devices by the `loopdev` container. When that happens,
the `disks-images-provider` pod enters the crash loop.

By detaching all loopback devices from the host at provider
startup we assure the next available device will be on range.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>