Stabilize SRIOV provider (#378)
* Ensure SRIOV-operator objects are created

By checking every X seconds Y times if a k8s-
object releated to sriov-operator is created,
instead of use sleep for arbitrary Z secondes.

Signed-off-by: Or Mergi <ormergi@redhat.com>

* Wait for sriov components pods to be ready

Sriov-cni and sriov-device-plugin are created after
new SriovNodeNetworkPolicy is created, it is necessary to
verify that tose pods are created and Ready before continuing.

Signed-off-by: Or Mergi <ormergi@redhat.com>

* Wait for 'NoSchedule' taint absence

After creating new SriovNetworkNodePolicy 'sriov-cni'
and 'sriov-device-plugin' pods are scheduled on the nodes
that match the policy, which cause a 'NoSchedule' taint on
that node that will remain untill those pods are up and ready.
Thus it is mandatory to wait for the taint to be removed
(after the mentioned pods are ready).

At 'wait_for_taint_absence' notice the '-v' flag at
the end of $action, propagates the string 'echo'ed
by is_taint_absence.

Signed-off-by: Or Mergi <ormergi@redhat.com>

* Wait for caBundle reconcile to finish

During caBundle reconcile 'NoSchedule' taint might placed on
the affected nodes, thuse in order to check if caBundle is
picked up, it is necessary to wait for the taint to
present and then wait for its absence.

Signed-off-by: Or Mergi <ormergi@redhat.com>

* Ensure VFs node resource are allocatable

It is nessacery to check if the node that
should be affected by the
SriovNodeNetworkPolicy has allocatable
resource that correspond to the policy.

Signed-off-by: Or Mergi <ormergi@redhat.com>

* Export multus deployment to function

Wait for multus damemonset ready replicas
('.status.numberReady') to equal desierd replicas
('.status.desiredNumberScheduled')

Signed-off-by: Or Mergi <ormergi@redhat.com>

* Move cert handaling to 'deploy_sriov_operator'

Now certificate patching to webhook-configurations
and cheking for taints after sriov-operator deployment
is in the right place.

Signed-off-by: Or Mergi <ormergi@redhat.com>

* Export apply sriov node policy code to a function

Signed-off-by: Or Mergi <ormergi@redhat.com>

* Add cluster is ready message

Signed-off-by: Or Mergi <ormergi@redhat.com>