Remove images.sh and adapt gocli to it
The k8s provider's digests SHAs are going to be stored at gocli binary so this commit
remove the images.sh and pass the SHAs at gocli Makefile so they are injected as build flags.

It also compiles cli binary if it's not already compiled at cluster-up to help with kubevirtci
dev iteration, in that case cli uses :latest.

Signed-off-by: Quique Llorente <ellorent@redhat.com>