Refactor calico on K8s 1.18 provider (#419)
* Align cni.yaml to original 3.12.0 yaml

In order to keep the original calico yaml,
patch it during provision and be able to
update it easily when needed, this commit
align the yaml to the official one.
Next commits would update the mechanisem of provision
and introduce the needed patch.

Signed-off-by: Or Shoval <oshoval@redhat.com>

* Update provision to use the original calico manifest

Provision updated to be based on the original calico
manifest, and then apply kubevirtci custom changes.
This way it would be clear what changes are applied,
and updating the manifest would be easier and safer.

Fixes: https://github.com/kubevirt/kubevirtci/issues/414

Signed-off-by: Or Shoval <oshoval@redhat.com>

* Update k8s-1.18 provider hash

Signed-off-by: Or Shoval <oshoval@redhat.com>