Bump calico to 3.15.1 for K8s-1.18 (#422)
* Bump calico to 3.15.1

In order to fix sonobuoy tests on dual stack,
a bump of calico is needed, because it updates
the CNI plugin to v0.8.6 which has an updated portmap
plugin.

Signed-off-by: Or Shoval <oshoval@redhat.com>

* Remove unneeded changes from calico cni diff file

Newer calico auto detect the
CALICO_IPV4POOL_CIDR and CALICO_IPV6POOL_CIDR
so manual update isnt needed.

From the docs:

Note: For clusters provisioned with kubeadm,
Calico autodetects the IPv4 and IPv6 pod CIDRs
and does not require configuration.

Signed-off-by: Or Shoval <oshoval@redhat.com>

* Update k8s-1.18 hash

Signed-off-by: Or Shoval <oshoval@redhat.com>