Configure custom IPv6 natted ippool
In order to allow IPv6 connectivity from Pods to the outside network, we
need to NAT the traffic. This patch configures calico CNI to do so.

This also changes the IPv6 CIDR of the internal cluster, so it does not
conflict with the external one. Without this change, routing does not
work as expected.

Signed-off-by: Petr Horacek <phoracek@redhat.com>