GoCli: Expose ports 80 and 443
When launching the cluster to be used as a testing environment it is
sometimes necessary to access the services inside the cluster through an
ingress configuration.
This change allows gocli to expose http and https and redirect any
request to node01 which, if configured as ingress controller, allow
access to the services in the cluster.

Signed-off-by: Gabriele Cerami <gcerami@redhat.com>