kind common, untangle kind config yaml preparation
On kind cluster creation We pass yaml file that represents
the cluster we want to create.
The file is located at kind/manifests/kind.yaml we
make a copy, do the changes and passing it tp kind binary
At the current state it is not simple to add more
changes to kind config yaml file, and its hard to predict
how the file will look like at the end.

This commit intent is to untange the way we change kind-config.yaml
by doing all the changes form one place, making it easier to maintain.

- '_prepare_kind_config()'
  Handels the logic for prepraring kind config yaml file
  and prints the final config for better visability.
- '_add_workers()'
  Iterates '$KUBEVIRT_NUM_NODES-1' and for each iteration
  and appends all necessary configuration for each worker.
- _add_worker_kubeadm_config_patch
  Appends kubeadmConfigPatch for a worker node

Signed-off-by: Or Mergi <ormergi@redhat.com>