sriov, Fix certsecret logic in case of etcd timeout
In case etcdserver: request timed out error occurs during Secret creation,
the secret might got created, and only the response was etcdserver: request timed out.
Current logic will Poll and fail repeatedly, ending with an error,
even that the secret was actually created.

Therefore the changes are:

* First make sure the Secret doesn't exists, in case it exists exit with error,
as certsecret isn't meant to be ran twice.
* In case there is an error when creating the secret, exit with success if the error is AlreadyExists,
because we know the secret didn't exist before we created it (see previous bullet).

Signed-off-by: Or Shoval <oshoval@redhat.com>