KIND infra, SRIOV proivder,  Run in-memory etcd (#478)
* kind common, untangle kind config yaml preparation

On kind cluster creation We pass yaml file that represents
the cluster we want to create.
The file is located at kind/manifests/kind.yaml we
make a copy, do the changes and passing it tp kind binary
At the current state it is not simple to add more
changes to kind config yaml file, and its hard to predict
how the file will look like at the end.

This commit intent is to untange the way we change kind-config.yaml
by doing all the changes form one place, making it easier to maintain.

- '_prepare_kind_config()'
  Handels the logic for prepraring kind config yaml file
  and prints the final config for better visability.
- '_add_workers()'
  Iterates '$KUBEVIRT_NUM_NODES-1' and for each iteration
  and appends all necessary configuration for each worker.
- _add_worker_kubeadm_config_patch
  Appends kubeadmConfigPatch for a worker node

Signed-off-by: Or Mergi <ormergi@redhat.com>

* sriov provider.sh, move kind config changes to kind/common.sh

Make use of 'kind_up'

Signed-off-by: Or Mergi <ormergi@redhat.com>

* kind infra, run etcd in memory

Currently we encounter bad performance of etcd on
sriov provider cluster on DinD setup.
We get 'etcdserver: timeout errors' that causes jobs to
fail often.
In cases where etcd has bad performance and the data shouldnt
be persistant (e.g: on CI and dev environments) it is recommanded [1]
to use in-memory etcd

To do that this commit:
-  Adds kubeadm ClusterConfiguration to kind config, setting
   etcd data directory to '/tmp/...' inside kind cluster nodes.
   '/tmp/' directory is already mounted to RAM memory as tmpfs.

- 'KUBEVIRT_WITH_KIND_ETCD_IN_MEMORY', expected variables: "true", "false"
   controls running etcd in memory on kind providers.

- 'ETD_DATA_DIR', expects directory path that mounted to RAM memory
   controls the path of etcd data directory inside kind cluster nodes

Running etcd in memory should improve performance and
will stabilize sriov provider and lanes.

[1] https://github.com/kubernetes-sigs/kind/issues/1922

Signed-off-by: Or Mergi <ormergi@redhat.com>