sriov, Fix certsecret backward compatibility
Use local imports so we can run certsecret
even on machines with older go (1.12.7 for example).

This method is also better, because it uses explicitly
the local certlib, and doesn't point to an external folder,
where the certlib might be different.

Without this change, updating go to 1.13.x was a must
Tried the following combination:
go1.12.14 to go1.13.15.

Signed-off-by: Or Shoval <oshoval@redhat.com>