config_sriov.sh: simplify 'deploy_sriov_operator'
After testing sriov-netwrok-operator deployment process
we found that it is possible to indicate that the
deploymnet is finished by waiting for all pods to be ready
after the deployment
Thus it is not necessary to wait for the taint to be
placed and removed  after deploymnet.

Also remove '|| return 1' and '|| exit 1' from
commands/functions function calls it is redundent
since we set bash flag '-e'.

Signed-off-by: Or Mergi <ormergi@redhat.com>