config_sriov.sh: simplify 'apply_sriov_node_policy'
After testing sriov deployment process for a while,
it is possible to inidcate that the sriov-operator is
ready after creating SriovNetworkNodePolicy by:
- Checking that all cluster pods are up and ready.
- Sriov capable nodes have sriov VF's as allocatable resource
  at the node status.

Thus, its no longer needed to wait for NoSchedule taint to be
placed and remove.
And no need to wait for 'sriov-cni' and 'sriov-devcie-plugin'
pods to create and become ready right after the policy is created.

Instead after creating SriovNetworkNodePolicy we will wait
for nodes status have sriov VFs as allocatable resource and
and for all cluster pods to be ready.

There is no need to use 'wait_k8s_object' and check that
SriovNetworkNodePolicy is created, if kubectl did not raised an
error the object it created.

Also remove '|| return 1' and '|| exit 1' from
commands/functions function calls it is redundent
since we set bash flag '-e'.

Signed-off-by: Or Mergi <ormergi@redhat.com>