SRIOV Provider fine tunes: Simplify Readiness Checks (#492)
* config_sriov.sh: refactor 'retry'

Change that condition of retry loop to stop when
'$action' return code is zero.
It will reduce 'retry' unexpected behaviour and will
make it easier to use (no need to hack the output)

Also extend 'retry' to run another action during the
waiting loop

Signed-off-by: Or Mergi <ormergi@redhat.com>

* config_sriov.sh: make 'wait_pods_ready' finite

At the current state 'wait_pod_ready' function would run infinitly
and due to pods that did not become ready, sriov lane prow jobs to
reached timeout and fail occasionally.

This commit put bounds to 'wait_pods_ready' using 'retry' and
presents 'check_all_pods_ready'.

Currently there is no simple kubectl command that returns all pods
that are not ready.
'check_all_pods_ready' checks each pod condition Ready,
this conditions will be true if all containers are ready and
all conditions specified in readinessGates are True [1].
Returns '0' if all pods conditions Ready equels 'true' otherwise '1'

[1] https://kubernetes.io/docs/concepts/workloads/pods/pod-lifecycle/#pod-readiness-status

Signed-off-by: Or Mergi <ormergi@redhat.com>

* config_sriov.sh: simplify 'deploy_sriov_operator'

After testing sriov-netwrok-operator deployment process
we found that it is possible to indicate that the
deploymnet is finished by waiting for all pods to be ready
after the deployment
Thus it is not necessary to wait for the taint to be
placed and removed  after deploymnet.

Also remove '|| return 1' and '|| exit 1' from
commands/functions function calls it is redundent
since we set bash flag '-e'.

Signed-off-by: Or Mergi <ormergi@redhat.com>

* config_sriov.sh: simplify 'apply_sriov_node_policy'

After testing sriov deployment process for a while,
it is possible to inidcate that the sriov-operator is
ready after creating SriovNetworkNodePolicy by:
- Checking that all cluster pods are up and ready.
- Sriov capable nodes have sriov VF's as allocatable resource
  at the node status.

Thus, its no longer needed to wait for NoSchedule taint to be
placed and remove.
And no need to wait for 'sriov-cni' and 'sriov-devcie-plugin'
pods to create and become ready right after the policy is created.

Instead after creating SriovNetworkNodePolicy we will wait
for nodes status have sriov VFs as allocatable resource and
and for all cluster pods to be ready.

There is no need to use 'wait_k8s_object' and check that
SriovNetworkNodePolicy is created, if kubectl did not raised an
error the object it created.

Also remove '|| return 1' and '|| exit 1' from
commands/functions function calls it is redundent
since we set bash flag '-e'.

Signed-off-by: Or Mergi <ormergi@redhat.com>

* config_sriov.sh: simplify 'deploy_multus'

There is no need to use 'wait_k8s_object' and check that
multus DaemonSet is created, if kubectl did not raised an
error the object it created.

Also remove '|| return 1' and '|| exit 1' from
commands/functions function calls it is redundent
since we set bash flag '-e'.

Signed-off-by: Or Mergi <ormergi@redhat.com>

* config_sriov.sh: remove conditional return oneliners at wait functions

Signed-off-by: Or Mergi <ormergi@redhat.com>

* config_sriov.sh: set '-e' flag explicitly

Currently '-e' flag is configured on kind/common.sh
that is sourced at the begining of this script.
Setting it in the config_sriov.sh makes the script
more self-contained.

Signed-off-by: Or Mergi <ormergi@redhat.com>