gocli, run command: support running etcd on RAM memory
In CI and dev environments where the cluster is ephemeral and
data shouldnt be persistant it is possible to improve cluster
performance by configuring etcd to run on RAM memory.
Hopfully it will shorten the time it takes for CI job to
finish and improve CI capacity.

This commit introduce new flag for gocli on run command
'run-etcd-on-memory' controls if etcd will run on RAM memory.

The path to etcd data direcory is configured at the manifest
being passed to kubeadm ('/var/lib/etcd').

Signed-off-by: Or Mergi <ormergi@redhat.com>