sriov, Assign only one PF for the worker
Currently sriov operator manifest actually uses just one of the PFs,
but allocates all available whitelisted PFs to the worker node.
Therefore this commit update logic
to allocate just one PF for the node, in default.

Modularity improvement:
Currently the PFs allocation function gets
the node it should allocate PFs to.
In order to increase modularity,
add a parameter which determines how many PFs it should allocate to the node.
This way we would be able to use it in a modular
way in case we need to control the number of PFs
we allocate to that node instead allocating
all the PFs to that node.

Signed-off-by: Or Shoval <oshoval@redhat.com>