sriov, Fix VFs allocation race condition of sriov operator
The flow is as follows:

1. The undercloud CNI of prow currently sets VFs number (sriov_numvfs)
to max allowed (according sriov_totalvfs).

2. The sriov operator is deployed.

3. The operator sets all unused PFs to have zero VFs,
since no policy applies to these PFs yet.
Number of VFs equal zero would eventually be reported at SriovNetworkNodeState.

4. Kubevirtci creates a new SriovNetworkNodePolicy.

5. The sriov operator checks if it should update number of VFs,
by comparing the desired numVfs (according to policy) to the current
numVfs configured according SriovNetworkNodeState status.

In case SriovNetworkNodeState didn't finish updating the actual
numVfs yet in the status,
the sriov operator would not create VFs, because the desired state
equals the current state.

The result would be that the cluster would stay with zero
VFs until a new policy is created,
or until the network config daemon is restarted.

By waiting for the SriovNetworkNodeState to finish updating
the status we fix this race condition.

Signed-off-by: Or Shoval <oshoval@redhat.com>