sriov, Assign only one PF for the worker (#510)
* sriov, Fix sriov pfs scan error detection

In case there are no sriov pfs available,
current sriov pfs scan logic would not detect it.
Bash would keep the array non empty,
and the only element would not be a valid pf
but the path itself.

Fix logic by using find, which returns an empty
output in case no valid matches found.

Since we use local keyword, the command would not
fail the script, even that we use -e.
Therefore add a check if the array is empty,
and fail fast in this case.
This will detect the problems as soon as possible
without trailing errors.

Signed-off-by: Or Shoval <oshoval@redhat.com>

* sriov, cosmetics of move_sriov_pfs_netns_to_node function

Rename ifs_name to pf_name as its actually a pf.

Add missing double quotes when using bash variables
in order to avoid word splitting, and in order to
comply with the other instances in the function.

Signed-off-by: Or Shoval <oshoval@redhat.com>

* sriov, Assign only one PF for the worker

Currently sriov operator manifest actually uses just one of the PFs,
but allocates all available whitelisted PFs to the worker node.
Therefore this commit update logic
to allocate just one PF for the node, in default.

Modularity improvement:
Currently the PFs allocation function gets
the node it should allocate PFs to.
In order to increase modularity,
add a parameter which determines how many PFs it should allocate to the node.
This way we would be able to use it in a modular
way in case we need to control the number of PFs
we allocate to that node instead allocating
all the PFs to that node.

Signed-off-by: Or Shoval <oshoval@redhat.com>