sriov, Improve CNI downloading (#531)
* sriov, Improve CNI downloading

Instead of downloading the CNI for every node,
download it once, and deploy it on all the nodes.

For local deployments it will just download it
if it doesnt exists already, which will improve
cluster-up timing more.

Signed-off-by: Or Shoval <oshoval@redhat.com>

* sriov, Rearrange installation of cnis in functions

Improve script structure and readability by
wrapping sections in functions according to their logic.

Signed-off-by: Or Shoval <oshoval@redhat.com>