sriov VM, Disable qemu guest agent at startup
There is a race condition at Kubevirt e2e tests,
between LoginToFedora and cloud-init which sets the password.

In order to fix this race condition and make sure
cloud-init finished, we disable quest agent when creating the VM image.
The e2e tests would enable it as the last thing
in cloud-init UserData, wait for the guest agent to finish
and only then login the the VM.

Note:
1. The next reboot won't be idempotent,
but atm we don't reboot sriov VMs in e2e tests.
2. In case NetworkData might happen after UserData
e2e should wait also for cloud-init finished indication.

See more info on
https://github.com/kubevirt/kubevirtci/pull/533
https://github.com/kubevirt/kubevirt/pull/4839#discussion_r561004638

Signed-off-by: Or Shoval <oshoval@redhat.com>