Move k8s 1.20 to crio CE (#541)
* Move k8s 1.20 to crio CE

Intent of this change is to switch from using docker container engine to crio.

There are two reasons to switch to crio:
docker deprecation in k8s
docker issue with selinux categories moby/moby#41370

Signed-off-by: Radim Hrazdil <rhrazdil@redhat.com>

* Use custom repository mirror

Signed-off-by: Radim Hrazdil <rhrazdil@redhat.com>

* Symlink docker to podman

Since docker is used outside kubevirtci, we still need to support it.
crictl however doesn't provide the same functiontality as docker.

In this commit, I'm linking docker to podman instead, which is designed
to have the same API.

Signed-off-by: Radim Hrazdil <rhrazdil@redhat.com>