conformance tests, sonobuoy: Control plugins to run from script
Currently each provider conformance config specifies
which sonobouy plugins [1] to run, which are e2e and
systemd-logs.

This commit changes from where we control which plugin
to run.
Instead of specifying the plugin name on each provider
conformance config, it will be controlled form the caller
script check-cluster-up.sh.

This will enable us to control which conformance test to run
and in what order.

[1] https://sonobuoy.io/docs/master/plugins/

Signed-off-by: Or Mergi <ormergi@redhat.com>