SRIOV provider: Add support for live migration
As part of Libvirt live migration it validates that
The source and target node has a unique product_uuid.

On containerized clusters (KIND infra) all nodes are actually
containers which have the same product_uuid.
In order to workaround this we creates PodPreset that mounts
fake random product_uuid to virt-launcher pods as presented
before [1].
PodPresets are deprecated on k8s-1.20 so it cant be used
for the long term, it's tracked on [2].

PodPrest is namespace scoped object,
Since we need to affect SRIOV VM's on Kubevirt sriov tests
the PodPreset should be created on the same namespace as the VM's,
Thus it's necessary to create the test namespaces in advance.

The PodPrest will be created on the namespace specified by
'SRIOV_TESTS_NS' variable.
'kubevirt-test-default1' is the default namespace of
Kubevirt SRIOV tests where the SRIOV VM's will be created.

This commit introduce podpreset.sh script to populate
necessary code for using PodPrest on KIND cluster.
podpreset.sh is aware of multi cluster setup.

[1] https://github.com/kubevirt/kubevirtci/pull/336
[2] https://github.com/kubevirt/kubevirtci/issues/571

Signed-off-by: Or Mergi <ormergi@redhat.com>