Make cluster-up/check.sh fail fast
Except for providers external or local we make the check script fail
fast. If the kvm device is not there or nested virtualization is not enabled
there is no chance the cluster will start up. In that case we make the
script exit with non zero error code to fail early.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>