docs, Update docs about testing locally provisioned clusters (#576)
* docs, Update CONTRIBUTING.md image

The gocli image is now hosted on quay.io
update docs accordingly.

Signed-off-by: Or Shoval <oshoval@redhat.com>

* docs, Rename KUBEVIRT_LOCAL_TESTING.md

Since the next commit going to refactor the file,
to include both kubevirtci and kubevirt notes
about locally running freshly provisioned kubevirtci
clusters, rename it accordingly.

Signed-off-by: Or Shoval <oshoval@redhat.com>

* docs, Update docs about testing locally provisioned clusters

This commit polish a bit the documentation on how to
test provisioned clusters.

* Add explanation on how to test gocli,
which is part of the provision toolchain.
* Update example to use a more recent VM based provider version
(k8s-1.19)
* Add needed exports for cluster-up to ease the reader.

Signed-off-by: Or Shoval <oshoval@redhat.com>