Istio, Add configuration for cluster without CNAO
Currently, Istio only works when CNAO is deployed at
the same time.

This commit adds Istio operator configuration
for the case when CNAO is not deployed, improving
the usability of istio in kubevirtci.

Additionally, the wait loop is improved to wait not
just for the istio-cni-node daemonset to be present
but also for all it's pods to be running.

Signed-off-by: Radim Hrazdil <rhrazdil@redhat.com>