Pre pull and check images for vanilla k8s test clusters (#407)
* Add fetch-images.sh, cdi manifests

Create a script to fetch all images required from either the provisioning scripts or the
manifests in the subfolder. Will be used to populate an image list file for pre-pulling
fixed version checking and after cluster-up checking actual deployed images.

Also add cdi manifests as they are a test dependency for kubevirt.

Signed-off-by: Daniel Hiller <daniel.hiller.1972@gmail.com>

* Extend fetch-images.sh, pre pull images, add check images

Add logic into fetch-images.sh to merge original image list with images
found in provisioning code.

Modify cli to inject image list. into container, so we can pre-pull during provisioning.

Add logic to check-cluster-up.sh for deploying the existing manifests and then fetching images
to compare them with list entries.

Signed-off-by: Daniel Hiller <daniel.hiller.1972@gmail.com>

* Add validation script for image list

Signed-off-by: Daniel Hiller <daniel.hiller.1972@gmail.com>

* Add pull policy validation for manifests

Create a small go tool that parses the manifests in the manifest
subdirectory in the provisioning folder and checks for either ex- or
implicit PullAlways policy.

Signed-off-by: Daniel Hiller <daniel.hiller.1972@gmail.com>

* Refactor and rename go tool

Signed-off-by: Daniel Hiller <daniel.hiller.1972@gmail.com>

* Add pod image pull policy check

Signed-off-by: Daniel Hiller <daniel.hiller.1972@gmail.com>

* Changed cdi operator manifest to v1.18.0

kubevirt/kubevirt is using the version when doing cluster-up for
testing, so in order to have these images pre-pulled we just use that
version now.

Signed-off-by: Daniel Hiller <daniel.hiller.1972@gmail.com>

* Change to using extra-pre-pull-images

Instead of using pre-pull-images as a list that contains all images that
should be pre pulled, we now use fetch-images.sh to calculate the images
that need to get pulled.

Signed-off-by: Daniel Hiller <daniel.hiller.1972@gmail.com>

* Add pre-pulling for 1.17

Apply logic changes for 1.17 k8s cluster.

Signed-off-by: Daniel Hiller <daniel.hiller.1972@gmail.com>

* Rename check image pull policies script

Rename the script to better reflect its intent. Also add some
explanation in the README.md to explain its usage.

Signed-off-by: Daniel Hiller <daniel.hiller.1972@gmail.com>

* Skip setting cnao setting for 1.14-1.16 clusters

Skipping to set the CNAO trigger setting for 1.14-1.16, also
remmoving the interactive and tty settings for docker run.
Signed-off-by: Daniel Hiller <daniel.hiller.1972@gmail.com>

* Use image for kubevirtci/check-image-pull-policies

Create Makefile for tools/check-image-pull-policies, add go fmt and
go mod tidy. Push image to kubevirtci repository and use that instead
of building the image every time.

Signed-off-by: Daniel Hiller <daniel.hiller.1972@gmail.com>

* Make checks for pull list explicit, add tests

Instead of warn on missing files, the tests for files required for image
pulling now fail the provisioning.

Also a set of tests was added to the go to for checking pull policies.

Signed-off-by: Daniel Hiller <daniel.hiller.1972@gmail.com>

* DNM: Update 1.21 with changes for pre pulling images

TODO: remove changes to 1.17 and 1.18 providers

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Revert changes to 1.17 and 1.18 provider provisioning

We only want the changes for the latest provider, so we are restoring
the initial setup from master again here.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Add missing images, skip already deployed manifests

Add all missing images to extra-pre-pull-images.

Also fix deployment to not include prometheus and ceph,
as those are already done during cluster-up. Update
filter for cni manifest name.

Increase pod check timeout for ci.

Fix error in fetch-images.sh that didn't extract images
in image@sha256:form, also exclude .svc image from istio

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Updated CDI to 1.34.1

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Added test and coverage targets

Signed-off-by: Daniel Hiller <dhiller@redhat.com>